Update Online_Store.sql
This project is about building the foundation (the database) for an online store. It involves:
1. Designing how the data will be organized.
2. Using MySQL to create that organization.
3. Using SQL to manage the data.
It doesn't include the website itself (the user interface) or the application logic that connects the website to the database, but it's a crucial part of making an online store work.

Database Schema Design:
My first step was to solidify the structure of my MySQL database. Based on the requirements, i used the following tables:
1. Products: Stores information about the items for sale.
2. Categories: Organizes products into different groups.
3. Users: Holds customer account details.
4. Addresses: Manages user shipping and billing addresses.
5. Orders: Records customer order information.
6. Order_items: Details the specific products included in each order.
7. Shopping_cart: Temporarily stores items in a user's cart.

After i have these tables set up, i then move on to populating them with sample data and then start writing queries to implement the online store's functionality.